finish userBook logic with merge and add recording

+++ src/modules/reading/application/use-case/user-book/delete.use-case.ts
@@ -3,6 +3,8 @@ import { type IUserBookRepository } from "@/modules/reading/domain/repositories/
 import { ObjectIdValueObject } from "@/shared/domain/value-object/object-id.vo";
 import { Inject, Injectable } from "@nestjs/common";
 import { USER_BOOK_REPOSITORY_TOKEN } from "@/modules/reading/domain/repositories/user-book.repository";
+import { UserBookEntity } from "@/modules/reading/domain/entities/user-book.entity";
+import { NotFoundException } from "@/shared/domain/exceptions/not-found.exception";
 
 /**
  * Delete user book use case
@@ -25,6 +27,13 @@ export class DeleteUserBookUseCase {
    * @returns True if the user book was deleted, false otherwise
    */
   public async execute(id: ObjectIdValueObject): Promise<boolean> {
+    // check if the user book exists
+    const userBook: UserBookEntity | null =
+      await this.userBookRepository.findById(id);
+    if (userBook === null) {
+      throw new NotFoundException("User book not found");
+    }
+
     // delete the user book
     return await this.userBookRepository.delete(id);
   }
